api-fara-docker-nou/users-microservice/src/main/java/com/socialsync/usersmicroservice/service/ValidateParameterService.java
package com.socialsync.usersmicroservice.service;

import org.springframework.stereotype.Service;

import java.util.regex.Pattern;
import java.util.regex.Matcher;

@Service
public class ValidateParameterService {
    public boolean isValidUsername(String s){
        String regex = "^[a-zA-Z0-9.]*$";
        Pattern pattern = Pattern.compile(regex);
        Matcher matcher = pattern.matcher(s);
        return matcher.matches();
    }

    public boolean isValidPassword(String s){
        if(s.length() < 5){
            return false;
        }
        String regex = "^[a-zA-Z0-9.]*$";
        Pattern pattern = Pattern.compile(regex);
        Matcher matcher = pattern.matcher(s);
        return matcher.matches();
    }

    public  boolean isValidEmail(String email) {
        String regex = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Z|a-z]{2,}$";
        Pattern pattern = Pattern.compile(regex);
        Matcher matcher = pattern.matcher(email);
        return matcher.matches();
    }
}
